+++ src/components/service/DatetimeWeather.vue
@@ -1,0 +1,69 @@
+<template>
+  <ServiceBase v-bind="props">
+    <template #icon="{ service }">
+      <ServiceBaseIcon :name="`wi:owm-${service?.data?.iconId}`" v-bind="iconProps" />
+    </template>
+    <template #title>
+      {{ currentDate }} | {{ currentTime }}
+    </template>
+    <template #description="{ service }">
+      {{ service?.data?.temp?.toFixed(1) }} {{ metricSymbol }} - {{ service?.data?.description }}
+    </template>
+  </ServiceBase>
+</template>
+
+<script setup lang="ts">
+import type { DatetimeWeatherService, ServiceClient } from '~/types'
+
+const props = defineProps<ServiceClient<DatetimeWeatherService>>()
+
+const { $settings } = useNuxtApp()
+const lang = computed(() => $settings.lang || 'en')
+const hour12 = computed(() => props.options?.timeFormat === '12h')
+const dateFormat = computed(() => props.options?.dateFormat)
+const timezone = computed(() => props.options?.timezone)
+
+const { formatDate } = useDateFormat(dateFormat, lang, timezone)
+
+const iconProps = computed(() => {
+  if (!props.icon) {
+    return {}
+  }
+  const { name: _, ...p } = props.icon
+  return p
+})
+
+const metricSymbol = computed(() => {
+  if (props?.options?.units === 'imperial') {
+    return '°F'
+  }
+  return '°C'
+})
+
+const now = ref(new Date())
+let timer: ReturnType<typeof setInterval> | undefined
+
+const timeFormatter = computed(() =>
+  new Intl.DateTimeFormat(lang.value, {
+    hour: '2-digit',
+    minute: '2-digit',
+    hour12: hour12.value,
+    timeZone: props.options?.timezone,
+  }),
+)
+
+const currentTime = computed(() => timeFormatter.value.format(now.value))
+const currentDate = computed(() => formatDate(now.value))
+
+onMounted(() => {
+  timer = setInterval(() => {
+    now.value = new Date()
+  }, 1000)
+})
+
+onBeforeUnmount(() => {
+  if (timer) {
+    clearInterval(timer)
+  }
+})
+</script>